enhancement: added contact info/comment field for each tunnel connection.
Allows for a short comment or an email address so you can remember who to contact for each tunnel.
The field is optional and does not need to be filled out to get a tunnel connection up.

Also cleaned up some of the layout on the tunnel pages and hopefully made it a bit easier to tell what is what.
Fixes #457