enhancement: add throughput info for each link to sysinfo.json
when `sysinfo.json` is called with `link_info=1` throughput info will be added for each link.
This runs grep against `/proc/net/dev` for each _linked_ interface over a 1 second interval.
it does slow down the response time of sysinfo.json, depending on how many linked interfaces there are.
I have many active tunnels and have not had any issues.

output is a string "RX TX" in **bytes** for that 1 second interval.

This could be useful for local graphing or other things.
I have some ideas to use it for the meshmap in the future.